[ENH] Refactor BOLD workflows
Create a `bold` module that makes more manageable contributing
to FMRIPREP.
