+++ fmriprep/workflows/bold/__init__.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+# emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
+# vi: set ft=python sts=4 ts=4 sw=4 et:
+# pylint: disable=unused-import
+
+from .base import init_func_preproc_wf
+from .hmc import init_bold_hmc_wf
+from .stc import init_bold_stc_wf
+from .registration import init_bold_reg_wf
+from .resampling import (
+    init_bold_mni_trans_wf,
+    init_bold_surf_wf,
+)
+
+from .confounds import (
+    init_bold_confs_wf,
+    init_ica_aroma_wf,
+)
