+++ fmriprep/workflows/bold/confounds.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 # emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
 # vi: set ft=python sts=4 ts=4 sw=4 et:
@@ -14,7 +13,7 @@ from niworkflows.nipype.algorithms import confounds as nac
 from niworkflows.interfaces import segmentation as nws
 from niworkflows.interfaces.masks import ACompCorRPT, TCompCorRPT
 
-from ..interfaces import (
+from ...interfaces import (
     TPM2ROI, ConcatROIs, CombineROIs, AddTSVHeader, GatherConfounds, ICAConfounds
 )
 
@@ -49,7 +48,7 @@ def init_bold_confs_wf(bold_file_size_gb, use_aroma, ignore_aroma_err, metadata,
         :graph2use: orig
         :simpleform: yes
 
-        from fmriprep.workflows.confounds import init_bold_confs_wf
+        from fmriprep.workflows.bold.confounds import init_bold_confs_wf
         wf = init_bold_confs_wf(bold_file_size_gb=1,
                                 use_aroma=True,
                                 ignore_aroma_err=True,
@@ -265,7 +264,7 @@ def init_ica_aroma_wf(name='ica_aroma_wf', ignore_aroma_err=False):
         :graph2use: orig
         :simpleform: yes
 
-        from fmriprep.workflows.confounds import init_ica_aroma_wf
+        from fmriprep.workflows.bold.confounds import init_ica_aroma_wf
         wf = init_ica_aroma_wf()
 
     **Parameters**
