+++ fmriprep/workflows/fieldmap/__init__.py
@@ -59,5 +59,6 @@ further details.
 
 """
 
-from .base import init_fmap_estimator_wf
-from .unwarp import init_sdc_unwarp_wf
+from .base import init_fmap_estimator_wf, init_fmap_unwarp_report_wf
+from .unwarp import init_sdc_unwarp_wf, init_pepolar_unwarp_wf
+from .syn import init_nonlinear_sdc_wf
